split confirmation page by feature
