+++ src/applications/check-in/#TODO:.md
@@ -1,9 +1,9 @@
 # TODO
 
 - display selected appointment
-  - behind feature toggle
-  - add time to header
-  - bold name
   - e2e feature toggle
   - e2e test
 - Organize imports
+  - ever file touched
+  organize props
+  - ever file touched
